sql/inspect: move inspect job entry point functions to inspect package
Move pkg/sql/inspect_job.go to pkg/sql/inspect/inspect_job_entry.go,
consolidating all inspect-related code in the inspect package.

Previously, functions for creating inspect jobs (TriggerInspectJob,
InspectChecksForTable, etc.) lived in pkg/sql. That package is already
quite big, so any chance to remove code from it is preferred.

Release note: None

Epic: None
