+++ pkg/sql/BUILD.bazel
@@ -141,7 +141,6 @@ go_library(
         "information_schema.go",
         "insert.go",
         "insert_fast_path.go",
-        "inspect_job.go",
         "instrumentation.go",
         "internal.go",
         "internal_result_channel.go",
